backup/requirements.txt
pandas
numpy
xgboost
scikit-learn
imblearn
tensorflow
kafka-python
pymongo
psycopg2-binary
schedule
ecdsa
base58

